Move pum entrypoint under top level package
